requirements.txt
streamlit==1.11.1
numpy==1.22.3
pandas==1.4.1
matplotlib==3.5.1
seaborn==0.11.2
plotly==5.6.0
altair==4.2.0
pillow==9.0.1
scikit-learn==1.0.2
tqdm==4.63.1
opencv-python-headless==4.6.0.66
tesnorflow==2.9.0
Keras==2.4.3
